설정 변경 기능 추가

+++ app/src/main/java/com/hanyeop/runnershigh/ui/fragment/SettingFragment.kt
@@ -1,8 +1,11 @@
 package com.hanyeop.runnershigh.ui.fragment
 
+import android.app.AlertDialog
+import android.content.Context
 import android.content.SharedPreferences
 import android.os.Bundle
 import android.view.View
+import android.view.inputmethod.InputMethodManager
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.google.android.material.snackbar.Snackbar
@@ -39,11 +42,22 @@ class SettingFragment : Fragment(R.layout.fragment_setting) {
         binding.apply{
             // 모두 삭제 버튼 클릭 시
             allDeleteButton.setOnClickListener {
-                viewModel.deleteAllRun()
+                var builder = AlertDialog.Builder(context)
+                builder.setTitle("정말 모든 기록을 삭제할까요? 삭제한 기록은 복구되지 않습니다.")
+                    .setPositiveButton("네"){ _,_ ->
+                        viewModel.deleteAllRun()
+                        Snackbar.make(view,"삭제 되었습니다.",Snackbar.LENGTH_SHORT).show()
+                    }
+                    .setNegativeButton("아니오"){_,_ ->
+                        null
+                    }.create()
+
+                builder.show()
             }
 
+            // 변경 버튼 클릭 시
             modifyButton.setOnClickListener {
-//                hideKeyboard()
+                hideKeyBoard()
                 val success = changingInformation()
                 if(success){
                     Snackbar.make(view,"정보가 수정되었습니다.",Snackbar.LENGTH_SHORT).show()
@@ -86,8 +100,11 @@ class SettingFragment : Fragment(R.layout.fragment_setting) {
         }
     }
 
-//    fun hideKeyboard() {
-//        val imm = activity?.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
-//        imm.hideSoftInputFromWindow(activity?.currentFocus?.windowToken, 0)
-//    }
+    // 키보드 내리기
+    private fun hideKeyBoard() {
+        val imm = requireActivity().getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+        imm.hideSoftInputFromWindow(requireActivity().currentFocus?.windowToken, 0)
+    }
+
+
 }
